+++ app/src/client/data-explorers/explorers/article-data/components/ArticleDataBySlugExplorer/stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import { InitialArticleDataBySlugPageData } from '@nlpssa-app-types/common/main';
 import storyData from 'client/_story-data';
-import { ArticleDataBySlugExplorer } from 'client/data-explorers/views/article-data/components';
+import { ArticleDataBySlugExplorer } from 'client/data-explorers/explorers/article-data/components';
 import { AppStateProvider } from 'client/data-explorers/store';
 
 const meta = {
