+++ app/src/client/data-explorers/explorers/article-data/components/index.ts
@@ -1,4 +1,4 @@
-import ArticleDataBySlugExplorer from 'client/data-explorers/views/article-data/components/ArticleDataBySlugExplorer';
-import ArticleDataExplorer from 'client/data-explorers/views/article-data/components/ArticleDataExplorer';
+import ArticleDataBySlugExplorer from 'client/data-explorers/explorers/article-data/components/ArticleDataBySlugExplorer';
+import ArticleDataExplorer from 'client/data-explorers/explorers/article-data/components/ArticleDataExplorer';
 
 export { ArticleDataBySlugExplorer, ArticleDataExplorer };
